Fix catching lightning with the EA
This fixes a bug where lightning was not properly bottled in a weather
container.

When registering recipes lightning should be registered first, as the
environmental accumulator will check for the weather types in the order
they are registered.
Secondly, the current weather was not being deactivated when the weather
should change, only the new weather was activated. This does not work
for lightning as activating rain during lightning simply does nothing.