Change dark temple spawning algorithm
The position of all spawned dark temples is saved in NBT and we use this
information to check if there are any dark temples in a given range
around the spot we want to spawn a new dark temple. The new temple will
only spawn when it is far enough away from other temples, this solves
the problem of temples spawning too close to each other.

The spawning conditions for the room needed by the dark temple is also
relaxed: Instead of needing an air space of 3x9x3, it is enough to NOT
have any solid blocks inside that area.

A last change involved allowing the center of a dark temple to spawn on
a spot grass is growing (thus replacing the grass when spawning).