Fix rendering for the refactored grenades
An icon is added for "empty" grenades. The icon that is rendered for
non-empty grenades is gotten from the last grenade type that is put on
the grenade (note: this might change in the future).

When grenades were dropped, they were also not rendered correctly and
neither when they were thrown. This should now be fixed.

Along with this a small improvement was made to RenderThrowable to keep
into account not only metadata, but the complete ItemStack (this allows
us to also use NBT data to render throwables).