Add (temporary) recipes for the new grenades.
Added some basic recipes that will probably change in the future, but at
least it shows the system is working.

Recipes need to be registered the "vanilla way", using the
registerCraftingRecipes() callback in IGrenadeType. Grenades in the
input recipe can be of any type and the properties of those grenades
will automatically be transferred to the output grenade.