Fix user crashing when farting on server
This fixes users crashing due to an incorrect display name that was sent
with the fart packet, which caused the target player to be null.

I also added an extra check and print a warning should the user still
not exist (e.g. because they lost connection right after sending a
packet).