Fix white border in purifier.
The rendering of the purifier showed a white border when rendered
ingame. Looking at the cauldron, they added a row of pixels in the
middle of the top texture which seems to fix the problem.

Oddly enough though, this border doesn't seem to be rendered at
all in game, perhaps this is a minecraft bug?